apps/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.1
python-multipart>=0.0.6
python-jose>=3.3.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.0.0

